make it easier to toggle thread safety: 149
while you can technically replace a default connection with a performant thread safe connection, that dependency doesnt feel great because it means i need to maintain a test of easypost that the interface doesnt change of the custom connection object i made. in an attempt to satisfy the benefit of those who are being single threaded, but give multi-folks a way to not get bottlenecked by deafult, thoughts on something like this?